+++ packages/google_mobile_ads/lib/src/ad_containers.dart
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 import 'dart:async';
-import 'dart:io' show Platform;
 import 'dart:math';
 
 import 'package:flutter/foundation.dart';
@@ -143,40 +142,6 @@ class LoadAdError extends AdError {
   }
 }
 
-/// Location parameters that can be configured in an ad request.
-class LocationParams {
-  /// Location parameters that can be configured in an ad request.
-  const LocationParams({
-    required this.accuracy,
-    required this.longitude,
-    required this.latitude,
-    this.time,
-  });
-
-  /// The accuracy in meters.
-  final double accuracy;
-
-  /// The longitude in degrees.
-  final double longitude;
-
-  /// The latitude in degrees.
-  final double latitude;
-
-  /// The UTC time, in milliseconds since epoch (January 1, 1970).
-  ///
-  /// This is required on Android, and ignored on iOS.
-  final int? time;
-
-  @override
-  bool operator ==(Object other) {
-    return other is LocationParams &&
-        accuracy == other.accuracy &&
-        longitude == other.longitude &&
-        latitude == other.latitude &&
-        time == other.time;
-  }
-}
-
 /// Targeting info per the AdMob API.
 ///
 /// This class's properties mirror the native AdRequest API. See for example:
@@ -189,7 +154,6 @@ class AdRequest {
     this.neighboringContentUrls,
     this.nonPersonalizedAds,
     this.httpTimeoutMillis,
-    this.location,
     this.mediationExtrasIdentifier,
     this.extras,
   });
@@ -215,12 +179,6 @@ class AdRequest {
   ///
   /// This is only supported in Android. This value is ignored on iOS.
   final int? httpTimeoutMillis;
-
-  /// Location data.
-  ///
-  /// Used for mediation targeting purposes.
-  @Deprecated('Location is not used and will be deleted in a future release')
-  final LocationParams? location;
 
   /// String identifier used in providing mediation extras.
   ///
@@ -258,7 +216,6 @@ class AdManagerAdRequest extends AdRequest {
     bool? nonPersonalizedAds,
     int? httpTimeoutMillis,
     this.publisherProvidedId,
-    LocationParams? location,
     String? mediationExtrasIdentifier,
     Map<String, String>? extras,
   }) : super(
@@ -267,7 +224,6 @@ class AdManagerAdRequest extends AdRequest {
           neighboringContentUrls: neighboringContentUrls,
           nonPersonalizedAds: nonPersonalizedAds,
           httpTimeoutMillis: httpTimeoutMillis,
-          location: location,
           mediationExtrasIdentifier: mediationExtrasIdentifier,
           extras: extras,
         );
@@ -857,15 +813,6 @@ class BannerAd extends AdWithView {
   @override
   final BannerAdListener listener;
 
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/6300978111'
-      : 'ca-app-pub-3940256099942544/2934735716';
-
   @override
   Future<void> load() async {
     await instanceManager.loadBannerAd(this);
@@ -1034,15 +981,6 @@ class NativeAd extends AdWithView {
   /// Options to configure the native ad request.
   final NativeAdOptions? nativeAdOptions;
 
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/2247696110'
-      : 'ca-app-pub-3940256099942544/3986624511';
-
   @override
   Future<void> load() async {
     await instanceManager.loadNativeAd(this);
@@ -1069,15 +1007,6 @@ class InterstitialAd extends AdWithoutView {
 
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<InterstitialAd>? fullScreenContentCallback;
-
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/1033173712'
-      : 'ca-app-pub-3940256099942544/4411468910';
 
   /// Loads an [InterstitialAd] with the given [adUnitId] and [request].
   static Future<void> load({
@@ -1179,15 +1108,6 @@ class RewardedAd extends AdWithoutView {
 
   /// Callbacks for events that occur when attempting to load an ad.
   final RewardedAdLoadCallback rewardedAdLoadCallback;
-
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/5224354917'
-      : 'ca-app-pub-3940256099942544/1712485313';
 
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<RewardedAd>? fullScreenContentCallback;